requirements.txt
langgraph>=0.2.55
langchain-google-genai>=2.1.0
langchain-core>=0.3.30
google-genai>=1.0.0
tavily-python>=0.5.0
mem0ai>=0.1.29
supabase>=2.10.0
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-dotenv>=1.0.1
pydantic>=2.10.3
httpx>=0.28.1
linkedin-api>=2.1.1
posthog>=3.7.4
python-jose[cryptography]>=3.3.0
pytest>=8.3.4
pytest-asyncio>=0.24.0
